+++ gamgyul-front/src/components/common/BasicLayout/layout.style.jsx
@@ -1,8 +1,8 @@
 import styled from "styled-components";
-import { theme } from "../../style/theme";
+import { theme } from "../../../style/theme";
 
 /** home Layout (기본 크기) */
-export const HomeLayout = styled.div`
+export const BasicLayout = styled.div`
   padding: 0 0 100px 0;
   height: calc(100vh - 100px);
   width: 100%;
